+++ franka_example_controllers/include/franka_example_controllers/fr3/move_to_start_example_controller.hpp
@@ -1,4 +1,4 @@
-// Copyright (c) 2026 Franka Robotics GmbH
+// Copyright (c) 2023 Franka Robotics GmbH
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -20,18 +20,15 @@
 #include <Eigen/Eigen>
 #include <controller_interface/controller_interface.hpp>
 #include <rclcpp/rclcpp.hpp>
-#include <std_msgs/msg/bool.hpp>
 
-#include "motion_generator.hpp"
+#include "franka_example_controllers/motion_generator.hpp"
 
 using CallbackReturn = rclcpp_lifecycle::node_interfaces::LifecycleNodeInterface::CallbackReturn;
 
 namespace franka_example_controllers {
 
-enum class ControlPhase { MOVE_TO_START, MOVE_TO_COLLISION, RETREAT, FINISHED };
-
 /// The move to start example controller moves the robot into default pose.
-class SelfCollisionFR3DuoExampleController : public controller_interface::ControllerInterface {
+class MoveToStartExampleController : public controller_interface::ControllerInterface {
  public:
   using Vector7d = Eigen::Matrix<double, 7, 1>;
   [[nodiscard]] controller_interface::InterfaceConfiguration command_interface_configuration()
@@ -45,30 +42,16 @@ class SelfCollisionFR3DuoExampleController : public controller_interface::Contro
   CallbackReturn on_activate(const rclcpp_lifecycle::State& previous_state) override;
 
  private:
-  std::vector<std::string> robot_types_;
-  std::vector<std::string> arm_prefixes_;
+  std::string robot_type_;
   const int num_joints = 7;
-
-  std::vector<Vector7d> q_;
-  std::vector<Vector7d> dq_;
-  std::vector<Vector7d> dq_filtered_;
-
-  std::vector<Vector7d> q_start_;
-  std::vector<Vector7d> q_collision_;
-
+  Vector7d q_;
+  Vector7d q_goal_;
+  Vector7d dq_;
+  Vector7d dq_filtered_;
   Vector7d k_gains_;
   Vector7d d_gains_;
-
   rclcpp::Time start_time_;
-  std::vector<std::unique_ptr<MotionGenerator>> motion_generators_;
-
-  rclcpp::Subscription<std_msgs::msg::Bool>::SharedPtr collision_sub_;
-  rclcpp::Time last_collision_msg_time_;
-  bool collision_detected_;
-  ControlPhase phase_;
-
-  const double kSpeedMotionGenerators = 0.2;
-  std::string collision_topic_;
+  std::unique_ptr<MotionGenerator> motion_generator_;
 
   void updateJointStates();
 };