+++ franka_example_controllers/src/fr3/cartesian_elbow_example_controller.cpp
@@ -13,43 +13,45 @@
 // limitations under the License.
 
 #include <franka_example_controllers/default_robot_behavior_utils.hpp>
-#include <franka_example_controllers/elbow_example_controller.hpp>
+#include <franka_example_controllers/fr3/cartesian_elbow_example_controller.hpp>
 
 #include <cassert>
 #include <cmath>
 #include <exception>
+#include <rclcpp/logging.hpp>
 #include <string>
 
 namespace franka_example_controllers {
 
 controller_interface::InterfaceConfiguration
-ElbowExampleController::command_interface_configuration() const {
+CartesianElbowExampleController::command_interface_configuration() const {
   controller_interface::InterfaceConfiguration config;
   config.type = controller_interface::interface_configuration_type::INDIVIDUAL;
-  config.names = franka_cartesian_velocity_->get_command_interface_names();
+  config.names = franka_cartesian_pose_->get_command_interface_names();
 
   return config;
 }
 
-controller_interface::InterfaceConfiguration ElbowExampleController::state_interface_configuration()
-    const {
+controller_interface::InterfaceConfiguration
+CartesianElbowExampleController::state_interface_configuration() const {
   controller_interface::InterfaceConfiguration config;
   config.type = controller_interface::interface_configuration_type::INDIVIDUAL;
-  config.names = franka_cartesian_velocity_->get_state_interface_names();
-
+  config.names = franka_cartesian_pose_->get_state_interface_names();
   // add the robot time interface
   config.names.push_back(robot_type_ + "/robot_time");
-
   return config;
 }
 
-controller_interface::return_type ElbowExampleController::update(
+controller_interface::return_type CartesianElbowExampleController::update(
     const rclcpp::Time& /*time*/,
     const rclcpp::Duration& /*period*/) {
   robot_time_ = state_interfaces_.back().get_optional<double>().value();
 
   if (initialization_flag_) {
-    initial_elbow_configuration_ = franka_cartesian_velocity_->getCurrentElbowConfiguration();
+    // Get initial elbow configuration values
+    initial_elbow_configuration_ = franka_cartesian_pose_->getCurrentElbowConfiguration();
+    // Get the initial pose
+    initial_pose_configuration_ = franka_cartesian_pose_->getCurrentPoseMatrix();
     initial_robot_time_ = robot_time_;
     elapsed_time_ = 0.0;
     initialization_flag_ = false;
@@ -58,15 +60,11 @@ controller_interface::return_type ElbowExampleController::update(
   }
 
   double angle = M_PI / 15.0 * (1.0 - std::cos(M_PI / 5.0 * elapsed_time_));
-
-  Eigen::Vector3d cartesian_linear_velocity(0.0, 0.0, 0.0);
-  Eigen::Vector3d cartesian_angular_velocity(0.0, 0.0, 0.0);
-
+  std::array<double, 16> pose_command = initial_pose_configuration_;
   std::array<double, 2> elbow_command = {
       {initial_elbow_configuration_[0] + angle, initial_elbow_configuration_[1]}};
 
-  if (franka_cartesian_velocity_->setCommand(cartesian_linear_velocity, cartesian_angular_velocity,
-                                             elbow_command)) {
+  if (franka_cartesian_pose_->setCommand(pose_command, elbow_command)) {
     return controller_interface::return_type::OK;
   } else {
     RCLCPP_FATAL(get_node()->get_logger(),
@@ -75,23 +73,22 @@ controller_interface::return_type ElbowExampleController::update(
   }
 }
 
-CallbackReturn ElbowExampleController::on_init() {
+CallbackReturn CartesianElbowExampleController::on_init() {
   return CallbackReturn::SUCCESS;
 }
 
-CallbackReturn ElbowExampleController::on_configure(
+CallbackReturn CartesianElbowExampleController::on_configure(
     const rclcpp_lifecycle::State& /*previous_state*/) {
-  franka_cartesian_velocity_ =
-      std::make_unique<franka_semantic_components::FrankaCartesianVelocityInterface>(
-          franka_semantic_components::FrankaCartesianVelocityInterface(k_elbow_activated_));
+  franka_cartesian_pose_ =
+      std::make_unique<franka_semantic_components::FrankaCartesianPoseInterface>(
+          franka_semantic_components::FrankaCartesianPoseInterface(k_elbow_activated_));
 
+  auto request = DefaultRobotBehavior::getDefaultCollisionBehaviorRequest();
   auto client = get_node()->create_client<franka_msgs::srv::SetFullCollisionBehavior>(
       "service_server/set_full_collision_behavior");
-  auto request = DefaultRobotBehavior::getDefaultCollisionBehaviorRequest();
 
   auto future_result = client->async_send_request(request);
   future_result.wait_for(robot_utils::time_out);
-
   auto success = future_result.get();
   if (!success) {
     RCLCPP_FATAL(get_node()->get_logger(), "Failed to set default collision behavior.");
@@ -118,24 +115,24 @@ CallbackReturn ElbowExampleController::on_configure(
   return CallbackReturn::SUCCESS;
 }
 
-CallbackReturn ElbowExampleController::on_activate(
+CallbackReturn CartesianElbowExampleController::on_activate(
     const rclcpp_lifecycle::State& /*previous_state*/) {
-  franka_cartesian_velocity_->assign_loaned_command_interfaces(command_interfaces_);
-  franka_cartesian_velocity_->assign_loaned_state_interfaces(state_interfaces_);
-
   initialization_flag_ = true;
   elapsed_time_ = 0.0;
+  franka_cartesian_pose_->assign_loaned_command_interfaces(command_interfaces_);
+  franka_cartesian_pose_->assign_loaned_state_interfaces(state_interfaces_);
+
   return CallbackReturn::SUCCESS;
 }
 
-controller_interface::CallbackReturn ElbowExampleController::on_deactivate(
+controller_interface::CallbackReturn CartesianElbowExampleController::on_deactivate(
     const rclcpp_lifecycle::State& /*previous_state*/) {
-  franka_cartesian_velocity_->release_interfaces();
+  franka_cartesian_pose_->release_interfaces();
   return CallbackReturn::SUCCESS;
 }
 
 }  // namespace franka_example_controllers
 #include "pluginlib/class_list_macros.hpp"
 // NOLINTNEXTLINE
-PLUGINLIB_EXPORT_CLASS(franka_example_controllers::ElbowExampleController,
+PLUGINLIB_EXPORT_CLASS(franka_example_controllers::CartesianElbowExampleController,
                        controller_interface::ControllerInterface)