+++ franka_example_controllers/test/test_gravity_compensation_example.cpp
@@ -16,14 +16,14 @@
 #include <memory>
 #include <vector>
 
-#include "franka_example_controllers/gravity_compensation_example_controller.hpp"
+#include "franka_example_controllers/fr3/gravity_compensation_example_controller.hpp"
 #include "rclcpp/rclcpp.hpp"
 
+#include <ros2_control_test_assets/descriptions.hpp>
 #include "hardware_interface/handle.hpp"
 #include "hardware_interface/types/hardware_interface_type_values.hpp"
 #include "rclcpp/utilities.hpp"
 #include "rclcpp_lifecycle/node_interfaces/lifecycle_node_interface.hpp"
-#include <ros2_control_test_assets/descriptions.hpp>
 
 using hardware_interface::CommandInterface;
 using hardware_interface::HW_IF_EFFORT;
@@ -74,9 +74,11 @@ void TestGravityCompensationExample::TearDown() {
 }
 
 void TestGravityCompensationExample::SetUpController() {
-  const auto result = controller_->init("test_gravitiy_compensation_example",ros2_control_test_assets::minimal_robot_urdf,30,"", rclcpp::NodeOptions()
-        .allow_undeclared_parameters(true)
-        .automatically_declare_parameters_from_overrides(true));
+  const auto result = controller_->init("test_gravitiy_compensation_example",
+                                        ros2_control_test_assets::minimal_robot_urdf, 30, "",
+                                        rclcpp::NodeOptions()
+                                            .allow_undeclared_parameters(true)
+                                            .automatically_declare_parameters_from_overrides(true));
   ASSERT_EQ(result, controller_interface::return_type::OK);
   std::vector<LoanedCommandInterface> command_ifs;
   command_ifs.emplace_back(joint_1_pos_cmd_);