+++ CHANGELOG.rst
@@ -21,6 +21,8 @@ Requires libfranka >= 0.20.4 and franka_description >= 2.4.0 requires ROS 2 Jazz
 * feat: update to libfranka 0.20.4
 * chore: cleanup franka_bringup launch utils import
 * feat: add franka_mobile_sensors as optional package for TMR robots
+* feat: mobile fr3 duo example added
+* refactor: franka example controllers divided into robot types
 
 v3.2.0 (2026-01-15)
 -------------------